requirements.txt
# Core deep learning
torch>=2.1.0
torchvision>=0.16.0

# Data processing & augmentation
numpy>=1.24
Pillow>=9.5.0
opencv-python>=4.8.0
albumentations>=1.3.1
kornia>=0.7.0

# Visualization
matplotlib>=3.7.0

# Utilities
tqdm>=4.66.0
gitpython>=3.1.30

# Logging and experiment tracking (optional)
wandb>=0.15.0

# Distributed training utilities (if used)
typing_extensions>=4.5.0
packaging>=23.0
